Added States
- Created new GameState class
- Modified GameApp to no longer be virtual base class but concrete class
in its own right that now owns a GameState object that is overridden for
each game state
- New GameState base class, has a unique_ptr to another GameState that
is normally null but when state change is needed a new state can be
added here.  GameApp will check if current state has a new state waiting
and make the change.
- Modified PongApp to be new state PongState
- Added new menu state
- Changed event handling of quit events to allow use of dialogs to
confirm quit.
- Also moved WindowCreateParams functions out of GameApp directly to
WindowCreationParams struct